feat: Improvement in visual behavior and optimization in components to add transactions
